Updated tutorial dialog and timings in a tutorial levels. Properly renamed Tutorial levels.
