+++ Puzzling Forest/Assets/Scenes/Tutorials/Tutorial_05_TwoFoxes.unity
@@ -2032,6 +2032,10 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 3313891023134956989, guid: 8ce517c63ad4afb4da231894f860402e, type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: 14.699951
+      objectReference: {fileID: 0}
     - target: {fileID: 4447354277789092346, guid: 8ce517c63ad4afb4da231894f860402e, type: 3}
       propertyPath: m_AnchoredPosition.x
       value: -15
@@ -2943,7 +2947,7 @@ GameObject:
   - component: {fileID: 594084645}
   - component: {fileID: 594084644}
   m_Layer: 0
-  m_Name: Tutorial_04
+  m_Name: Tutorial_05
   m_TagString: ScriptTrigger
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
