+++ Puzzling Forest/Assets/Scripts/SceneManagers/TutorialScripts/Tutorial_04.cs
@@ -4,92 +4,46 @@ using UnityEngine;
 
 public class Tutorial_04 : TutorialScript
 {
-    private UnityEngine.UI.Button nextButton;
+	private Vector3 newLeftSide = new Vector3(1.42f, 1.29f, -4.56f);
+	private Vector3 newRightSide = new Vector3(-4.66f, 1.29f, 1.52f);
 
-    private string welcomeMessage = "Whoa! Now there's two foxes. We'll have to help them both to reach the house!";
 
-    private const int numSections = 2;
-    private int curSection = 0;
-    private string[] sections = new string[numSections] { "Welcome", "JustTips" };
-    private float timeAtLastPress = -1f;
+	protected override void Awake()
+	{
+		base.Awake();
+	}
 
-    private bool isFirstBlockDone = false;
+	// Start is called before the first frame update
+	protected override void Start()
+	{
+		base.Start();
 
-    // Start is called before the first frame update
-    protected override void Awake()
-    {
-        base.Awake();
+		StartCoroutine(Begin());
+	}
 
-        nextButton = FairyCanvas.transform.Find("Background/NextButton").GetComponent<UnityEngine.UI.Button>();
-    }
+	// Update is called once per frame
+	protected override void Update()
+	{
+		base.Update();
+	}
 
-    protected override void Start()
-    {
-        StartNextSection();
-    }
+	private IEnumerator Begin()
+	{
+		yield return new WaitForSeconds(0.1f);
+		FairyController = turnManager.GetCurrentFairy().GetComponent<IndicatorAnimationController>();
+		//FairyController.resizeCanvas(850f, 375f);
 
-    // Update is called once per frame
-    protected override void Update()
-    {
-        base.Update();
-    }
+		yield return new WaitForSeconds(turnManager.Say("This puzzle is a bit trickier.", typingNoise) - 0.1f);
 
-    public void StartNextSection()
-    {
-        if (Time.time - timeAtLastPress < 1f)
-            return;
-        else
-        {
-            timeAtLastPress = Time.time;
-            StartCoroutine(sections[curSection]);
-            curSection++;
-        }
-    }
+		yield return new WaitForSeconds(turnManager.Say("If you get stuck, you can press 'U' to undo your most recent move.", typingNoise) + 0.1f);
+		TipsCanvas.transform.Find("TipsMenu/Tip_01").gameObject.SetActive(true);
+		if (!isTipsShown)
+			ToggleTips();
+		alertNoise.Play();
+	}
 
-    private IEnumerator Welcome()
-    {
-        yield return rapidPause;
-        turnManager.StealControl();
-        timer.SetPause(true);
-
-        FairyAnchor.SetActive(true);
-        nextButton.interactable = false;
-
-        yield return StartCoroutine(Type(FairyText, welcomeMessage));
-        nextButton.interactable = true;
-    }
-
-    private IEnumerator JustTips()
-    {
-        ClearText(FairyText);
-
-        FairyController.FlyOff();
-        yield return flyPause;
-
-        FairyAnchor.SetActive(false);
-
-        turnManager.ResumeControl();
-        timer.SetPause(false);
-
-        yield return new WaitForSeconds(turnManager.Say("Press 'E' to change which fox we're helping.", typingNoise) - 3f);
-
-        TipsCanvas.transform.Find("TipsMenu/Tip_01").gameObject.SetActive(true);
-        TipsCanvas.transform.Find("TipsMenu/Tip_02").gameObject.SetActive(true);
-        if (!isTipsShown)
-            ToggleTips();
-        alertNoise.Play();
-    }    
-
-    void OnTriggerEnter(Collider other)
-    {
-        if (!isFirstBlockDone)
-        {
-            isFirstBlockDone = true;
-            turnManager.Say("One down, one to go.", typingNoise);
-        }
-        else
-        {
-            turnManager.Say("Great job.", typingNoise);
-        }
-    }
+	void OnTriggerEnter(Collider other)
+	{
+		turnManager.Say("Great job!", typingNoise);
+	}
 }
