+++ Puzzling Forest/ProjectSettings/EditorBuildSettings.asset
@@ -11,9 +11,6 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/Menus/Level Select.unity
     guid: e2d5db07b51a74c4b8834d6ce2963cf5
-  - enabled: 0
-    path: 
-    guid: 00000000000000000000000000000000
   - enabled: 1
     path: Assets/Scenes/z_CutScenes/Intro.unity
     guid: 929838cf43052b34382ce6d1a7f0fd4f
@@ -27,10 +24,13 @@ EditorBuildSettings:
     path: Assets/Scenes/Tutorials/Tutorial_02_BlockPuzzles.unity
     guid: fe1b54a31cff9e14387429eac177d44c
   - enabled: 1
-    path: Assets/Scenes/Tutorials/Tutorial_03_MultiblockPuzzles.unity
+    path: Assets/Scenes/Tutorials/Tutorial_03_MultiblockPushing.unity
+    guid: 2afba5d96c80a88459c9eab7bb545cfe
+  - enabled: 1
+    path: Assets/Scenes/Tutorials/Tutorial_04_MultiblockPuzzles.unity
     guid: 1332b9a87d6db674da179d2acee0b6f3
   - enabled: 1
-    path: Assets/Scenes/Tutorials/Tutorial_04_TwoFoxes.unity
+    path: Assets/Scenes/Tutorials/Tutorial_05_TwoFoxes.unity
     guid: aaea9fa6bc1620e4a83da7cfd3243256
   - enabled: 1
     path: Assets/Scenes/Garrison_1.unity
@@ -39,7 +39,7 @@ EditorBuildSettings:
     path: Assets/Scenes/Leon_Level_17.unity
     guid: 70d614f712c00f848befbd1c88421660
   - enabled: 1
-    path: Assets/Scenes/Tutorials/Garrison_2.unity
+    path: Assets/Scenes/Tutorials/Tutorial_06_Foxholes.unity
     guid: 1f5795b5c2b689e4cbcf27749ce2a78b
   - enabled: 1
     path: Assets/Scenes/Leon_Level_1.unity
@@ -50,9 +50,6 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/Nathaniel Level 1.unity
     guid: 20335ed10d92f8044b236d02b87f2823
-  - enabled: 0
-    path: Assets/Scenes/Tutorials/Wyatt_Level_1.unity
-    guid: 2afba5d96c80a88459c9eab7bb545cfe
   - enabled: 0
     path: Assets/Scenes/Justin_Level_036.unity
     guid: 5fa954672bda33741900f7fe6014b12f
@@ -152,9 +149,6 @@ EditorBuildSettings:
   - enabled: 0
     path: Assets/Scenes/RobertScene10.unity
     guid: e24fe7491b628a346a0a017f14669a2f
-  - enabled: 0
-    path: 
-    guid: 00000000000000000000000000000000
   - enabled: 0
     path: Assets/Scenes/RobertScene11.unity
     guid: c4338f9f6af13bf47b37823cdd858983
